Assicurata la creazione della directory target

+++ build_compose.sh
@@ -42,11 +42,8 @@ while getopts "shv:" opt; do
   esac
 done
 
-if [ -d ./target ]
-then
-   rm -rf ./target
-   mkdir ./target
-fi
+[ -d ./target ] && rm -rf ./target
+mkdir ./target
 
 cp -rp ./commons/resources_compose ./commons/catalina_wrapper.sh ./target
 if [ -z "$DA_SORGENTI" ]
